+++ src/app/shared/fhir-backend/fhir-batch-query.js
@@ -63,7 +63,7 @@ class FhirBatchQuery extends EventTarget {
     super();
     this._serviceBaseUrl = serviceBaseUrl;
     this._authorizationHeader = null;
-    this._scrubberID = null;
+    this._scrubberID = {};
     this._pending = [];
     this._batchTimeoutId = null;
     this._batchTimeout = batchTimeout || globalThis.sessionStorage?.getItem('batchTimeout') || 20;
@@ -107,18 +107,20 @@ class FhirBatchQuery extends EventTarget {
 
   /**
    * Returns the scrubber ID header value.
-   * @return {string} - the scrubber ID.
+   * @param {string|null} [serviceBaseUrl] - the FHIR REST API Service Base URL
+   * @return {string|undefined} - the scrubber ID.
    */
-  getScrubberIDHeader() {
-    return this._scrubberID;
+  getScrubberIDHeader(serviceBaseUrl = null) {
+    return this._scrubberID[serviceBaseUrl || this._serviceBaseUrl];
   }
 
   /**
    * Sets the scrubber ID header value.
    * @param {string} value - the scrubber ID.
+   * @param {string|null} [serviceBaseUrl] - the FHIR REST API Service Base URL
    */
-  setScrubberIDHeader(value) {
-    this._scrubberID = value;
+  setScrubberIDHeader(value, serviceBaseUrl = null) {
+    this._scrubberID[serviceBaseUrl || this._serviceBaseUrl] = value;
     FhirBatchQuery.clearCache();
   }
 
@@ -846,8 +848,9 @@ class FhirBatchQuery extends EventTarget {
         oReq.setRequestHeader('Authorization', this._authorizationHeader);
       }
 
-      if (this._scrubberID) {
-        oReq.setRequestHeader('Scrubber-Id', this._scrubberID);
+      const scrubberId = this.getScrubberIDHeader()
+      if (scrubberId) {
+        oReq.setRequestHeader('Scrubber-Id', scrubberId);
       }
 
       if (this.withCredentials) {